Consolidate paywall cache warming tests
Keep traversal and cache-consumer coverage under the existing cache warming test class so all supported project configurations discover the tests.
